http-trigger/requirements.txt
google-cloud-pubsub==2.13.0
google-cloud-firestore==2.7.2
functions-framework==3.0.0